Se agrega modulo Shop

+++ src/app/pages/admin/products/products.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FloatingActionButton } from 'ng-floating-action-menu';
 import { CategoriesService } from 'src/app/services/core/categories.service';
 import { ProductsService } from 'src/app/services/core/products.service';
+import Swal from 'sweetalert2'
+import { Router } from '@angular/router';
 
 @Component({
 	selector: 'app-products',
@@ -16,12 +18,11 @@ export class ProductsComponent implements OnInit {
 	defaultBindingsList = [];
 	selectedBrand = null;
 
-	
-
 	constructor(private productsSvc: ProductsService, 
-				private categoriesSvc: CategoriesService) {
+				private categoriesSvc: CategoriesService,
+				private router: Router) {
+
 		this.getAllProducts();
-	
 		this.categoriesSvc.getCategoriesSelect().subscribe((data: any) => {
 			this.defaultBindingsList = data.listCategories;
 		});
@@ -62,11 +63,35 @@ export class ProductsComponent implements OnInit {
 
 	}
 
+	click_addProduct(){
+		this.router.navigateByUrl('/admin/products/new');
+	}
+
 	click_deleteProduct(id:number){
-		this.productsSvc.deleteProduct(id).subscribe((data:any)=>{
-			console.log(data);
-			this.getAllProducts();
+		Swal.fire({
+			title: 'Estas seguro?',
+			text: "Ya no podrás revertir esto!",
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonText: 'Si, eliminar!',
+			cancelButtonText: 'No, cancelar!',
+			confirmButtonColor: '#6a1b9a',
+			cancelButtonColor:'#CC0000',
+			reverseButtons: true
+		}).then((result) => {	
+			if(result.value){
+				this.productsSvc.deleteProduct(id).subscribe((data:any)=>{
+					this.getAllProducts();
+				});
+			}else{
+				Swal.fire(
+					'Cancelado',
+					'El Item se encuentra seguro =)',
+					'error'
+				);
+			}
 		});
+			
 		
 	}
